apps/image-recognition-training/requirements.txt
pillow==10.3.0
requests==2.32.3
torch==2.3.1
transformers==4.41.2